requirements.txt
numpy
av
Pillow
cv2
